atualização para responsividade e novas fotos
