--- src/paginas/2-Turmas/TurmaAtual.js
@@ -1,10 +1,0 @@
-function TurmaAtual(){
-    return(
-        <div>
-            <h1>Turma Atual</h1>
-            <p>Conteúdo da página</p>
-        </div>
-    )
-}
-
-export default TurmaAtual
