--- src/paginas/2-Turmas/TurmasAnteriores.css
@@ -1,111 +1,0 @@
-.carousel-turmas {
-  width: 90%;
-  margin: auto;
-  text-align: center;
-}
-
-.carousel-turmas .aluno {
-  cursor: grab;
-  height: 380px;
-  color: black;
-  box-sizing: border-box;
-}
-
-.carousel-turmas .aluno p {
-  text-align: center !important;
-  max-width: 180px;
-  margin: auto;
-  font-family: 'Trebuchet MS', 'Lucida Sans Unicode', 'Lucida Grande', 'Lucida Sans', Arial, sans-serif;
-  font-size: 1.1vw;
-}
-
-.carousel-turmas .aluno img {
-  padding: 10px;
-  height: 300px;
-  width: 200px;
-  margin: auto;
-  border-radius: 30px;
-}
-
-.slick-dots li button:before {
-  font-size: 20px;
-  color: #10813c;
-  padding-top: 30px;
-  cursor: pointer;
-}
-
-.slick-dots li.slick-active button:before {
-  color: #10813c;
-}
-
-.buttons-turmas {
-  margin-bottom: 30px;
-}
-
-.buttons-turmas * {
-  display: inline-block;
-  margin: 0 3vw;
-  padding: 1.2vw 2.4vw;
-  background-color: #10813c;
-  color: #fff;
-  border: none;
-  border-radius: 5px;
-  font-size: 16px;
-  font-weight: bold;
-  text-decoration: none;
-  text-align: center;
-  transition: background-color 0.3s ease;
-  cursor: pointer;
-}
-
-.buttons-turmas *:hover {
-  background-color: #0f3100;
-}
-
-.buttons-turmas .selected {
-  background-color: #0f3100;
-}
-
-.slider-default {
-  cursor: default !important;
-}
-
-@media (max-width: 1024px){
-  .carousel-turmas .aluno p{
-    font-size: 1.6vw;
-  }
-}
-
-@media (max-width: 1000px){
-  .carousel-turmas .aluno p{
-    font-size: 1.9vw;
-  }
-}
-
-@media (max-width: 700px){
-  .carousel-turmas .aluno p{
-    font-size: 3.5vw;
-  }
-  .buttons-turmas * {
-    font-size: 15px;
-    line-height: 25px;
-    padding: 10px 12px;
-  }
-}
-
-@media (max-width: 400px){
-  .carousel-turmas .aluno p{
-    font-size: 4.5vw;
-  }
-}
-
-@media (max-width: 300px){
-  .carousel-turmas .aluno p{
-    font-size: 5vw;
-  }
-  .buttons-turmas * {
-    font-size: 13px;
-    line-height: 25px;
-    padding: 10px 12px;
-  }
-}
