+++ src/paginas/3-Equipe/equipe.css
@@ -37,7 +37,7 @@
 .equipe-container .card-equipe img {
     padding: 10px;
     height: 300px;
-    width: 200px;
+    width: 220px;
     margin: auto;
     border-radius: 30px;
 }
@@ -64,6 +64,21 @@
     text-align: center;
     transition: background-color 0.3s ease;
     cursor: pointer;
+}
+
+.select-equipe {
+    display: inline-block;
+    margin: 0 3vw;
+    padding: 4.2vw 8.4vw;
+    background-color: #10813c;
+    color: #fff;
+    border: none;
+    border-radius: 5px;
+    font-size: 16px;
+    font-weight: bold;
+    text-decoration: none;
+    text-align: center;
+    margin-bottom: 20px;
 }
 
 .buttons-equipe *:hover {
